requirements.txt
tkinter
pathlib
requests
beautifulsoup4
selenium
lxml
chardet
python-magic
pycryptodome
pyyaml 